backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "openai>=1.78.0",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
]
